Give deprecation warning if ifPresent or required are used on non-nullable properties
